Moved SessionEditPage.js to SessionPage folder and styled the code a little
